+++ CHANGELOG.md
@@ -5,3 +5,4 @@
 ### Changed
 
 - CollectionView: added default implementation for "getHTML()" method returning simple "div" element.
+- Documentation: improve example usage of "OsfObservable.retrigger()" method.