+++ CHANGELOG.md
@@ -7,3 +7,4 @@
 - CollectionView: added default implementation for "getHTML()" method returning simple "div" element.
 - Documentation: improve example usage of "OsfObservable.retrigger()" method.
 - Documentation: added "async" prefix to examples of View method usages
+- Documentation: fixed formatting in various code examples