+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - CollectionView: allowed not to specify EmptyView when it's not used
 - CollectionView: allowed to render Views inside of specified child Element
 - CollectionView: EmptyView is automatically removed on "addChildView()" calls
+- CollectionView: allowed to be instantiated without inheritance
 - Documentation: improve example usage of "OsfObservable.retrigger()" method
 - Documentation: added "async" prefix to examples of View method usages
 - Documentation: fixed formatting in various code examples